refact: execution separate and refact all executions
