+++ include/ast.h
@@ -6,7 +6,7 @@
 /*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/18 08:43:01 by rde-mour          #+#    #+#             */
-/*   Updated: 2024/04/14 15:43:26 by rde-mour         ###   ########.org.br   */
+/*   Updated: 2024/04/27 14:23:30 by rde-mour         ###   ########.org.br   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@ struct s_ast
 };
 
 void	ast_print(t_ast **ast);
+void	ast_remove(t_ast **ast);
 t_ast	*ast_new(t_list **tokens);
 t_ast	*ast_node(t_list **tokens);
 t_ast	*ast_clear(t_ast **ast);
