+++ include/builtin.h
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   builtin.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/04/24 19:22:28 by rde-mour          #+#    #+#             */
+/*   Updated: 2024/04/27 21:33:38 by rde-mour         ###   ########.org.br   */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef BUILTIN_H
+# define BUILTIN_H
+
+# include "ast.h"
+# include "expansions.h"
+
+int	builtin_exit(t_env *env, t_ast **ast, char **cmd, int *fds);
+
+#endif
