+++ src/ast/ast.c
@@ -6,7 +6,7 @@
 /*   By: rde-mour <rde-mour@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/18 08:42:24 by rde-mour          #+#    #+#             */
-/*   Updated: 2024/04/20 21:03:16 by rde-mour         ###   ########.org.br   */
+/*   Updated: 2024/04/27 16:16:01 by rde-mour         ###   ########.org.br   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,32 @@
 #include "prompt.h"
 #include "parser.h"
 #include "ft_string.h"
+
+
+void	ast_print(t_ast **ast)
+{
+	t_ast	*tmp;
+
+	if (!ast || !(*ast))
+		return ;
+	tmp = *ast;
+	if ((*ast)->left)
+		ast_print(&(*ast)->left);
+	if ((*ast)->right)
+		ast_print(&(*ast)->right);
+	printf("%d %s\n", tmp->content->type, tmp->content->literal);
+}
+
+
+void	ast_remove(t_ast **ast)
+{
+	if (!ast || !(*ast))
+		return ;
+	free((*ast)->content->literal);
+	free((*ast)->content);
+	free(*ast);
+	*ast = NULL;
+}
 
 t_ast	*ast_clear(t_ast **ast)
 {
@@ -36,20 +62,6 @@ t_ast	*ast_clear(t_ast **ast)
 	free(tmp);
 	ast = NULL;
 	return (NULL);
-}
-
-void	ast_print(t_ast **ast)
-{
-	t_ast	*tmp;
-
-	if (!ast || !(*ast))
-		return ;
-	tmp = *ast;
-	if ((*ast)->left)
-		ast_print(&(*ast)->left);
-	if ((*ast)->right)
-		ast_print(&(*ast)->right);
-	printf("%d %s\n", tmp->content->type, tmp->content->literal);
 }
 
 static t_ast	*ast_get(t_ast **paren)
